repercute parsing fixes on bonus
