+++ bonus/srcs/parsing/get_textures_utils_bonus.c
@@ -92,16 +92,3 @@ int	get_textures_line(char *buff, char *line_array[4], int *data_got,
 	}
 	return (EXIT_SUCCESS);
 }
-
-int	check_dup(char *buff, t_tracker *tracker)
-{
-	if (ft_strncmp(buff, "NO", 2) == 0 && tracker->north)
-		return (ft_dprintf(2, ERM_DUPLI, "north"), true);
-	else if (ft_strncmp(buff, "SO", 2) == 0 && tracker->south)
-		return (ft_dprintf(2, ERM_DUPLI, "south"), true);
-	else if (ft_strncmp(buff, "WE", 2) == 0 && tracker->west)
-		return (ft_dprintf(2, ERM_DUPLI, "west"), true);
-	else if (ft_strncmp(buff, "EA", 2) == 0 && tracker->east)
-		return (ft_dprintf(2, ERM_DUPLI, "east"), true);
-	return (false);
-}
